Formatting changes and last minute fixes for web version. Detailed changelog below:
Fixed broken reference to [Wa2018] in intro.
Center-aligned the "landscape" figure.
Re-ran all cells in order.
Changed cirq installation to install the version that I'm using for the notebook. (For longevity).
Changed text to reflect this.
Changed comment of first code cell installing Cirq.
Changed code in cell 4 to adjust to margins.
Changed comment of In[6] to adjust to margins.
Fixed wrong print statement of In[7] by re-running all cells.
Changed comment of """Simulator perfromance test...""" for margins.
Changed comment of In[11] for margins.
Changed comment of In[14] for margins.
Changed code of In[14] for margins.
Changed code of In[19] for margins.
Changed code of In[21] for margins.
Fixed comment of In[29] for margins.
Fixed comment of In[32] for margins.
Changed code for In[33] to decrease the size of the print statement for margins.
Fixed incorrect output for optimizing the Bell state preparation circuit. (The text describes one circuit but I ran it for two circuits.)
Touched up code and decreased the length of lines in the last VQSD example.
Added reference to a recent quantum software paper in intro.
Fixed reference [Al2018] to be [Ho2018] for consistency.